[BEAM-7303] Rename runners.reference package to runners.portability
