+++ runners/portability/java/build.gradle
@@ -17,9 +17,9 @@
  */
 
 plugins { id 'org.apache.beam.module' }
-applyJavaNature(automaticModuleName: 'org.apache.beam.runners.reference')
+applyJavaNature(automaticModuleName: 'org.apache.beam.runners.portability')
 
-description = "Apache Beam :: Runners :: Reference :: Java"
+description = "Apache Beam :: Runners :: Portability :: Java"
 ext.summary = """A Java implementation of the Beam Model which utilizes the portability
 framework to execute user-definied functions."""
