+++ runners/portability/java/src/main/java/org/apache/beam/runners/portability/testing/TestPortablePipelineOptions.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.runners.reference.testing;
+package org.apache.beam.runners.portability.testing;
 
 import com.google.auto.service.AutoService;
 import org.apache.beam.runners.fnexecution.jobsubmission.JobServerDriver;
