+++ runners/portability/java/src/main/java/org/apache/beam/runners/portability/testing/TestPortableRunner.java
@@ -15,13 +15,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.runners.reference.testing;
+package org.apache.beam.runners.portability.testing;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 
 import java.io.IOException;
 import org.apache.beam.runners.fnexecution.jobsubmission.JobServerDriver;
-import org.apache.beam.runners.reference.PortableRunner;
+import org.apache.beam.runners.portability.PortableRunner;
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.PipelineResult;
 import org.apache.beam.sdk.PipelineResult.State;
