+++ settings.gradle
@@ -51,7 +51,7 @@ include ":runners:google-cloud-dataflow-java:examples-streaming"
 include ":runners:java-fn-execution"
 include ":runners:jet"
 include ":runners:local-java"
-include ":runners:reference:java"
+include ":runners:portability:java"
 include ":runners:spark"
 include ":runners:spark:job-server"
 include ":runners:samza"
